Added documentation and reorganized code structure
This commit adds a `docs` folder with a `GettingStarted.md` file, an `img` folder to the `docs` folder, and an `examples` folder. It also moves some files around: `consts.go`, `structs.go`, and `go-nyse-stocks.go` are moved to the new `src` folder. The README.md file is updated to reflect these changes.